Teleport Updates & Minor Fix
Added some more locations to teleport to via Radio Buttons.

Added a minor fix regarding calculating the "Code Term". Before it would
"roll over" to 0x02 if the code term went above 0xFF, but because it
increments by 3, that means 0xFE + 0x03 would need to have a code term
of 0x01, not 0x02. Included 3 if statements for 0xFD, 0xFE and 0xFF.