Major XOR calculation overhaul
Changed the way the XOR values are calculated (the user no longer needs
to input their in-game strength value as a base for calculations). Most
of the functions used before have been commented out and may be removed
in a later version. Teleportation tool is STILL broken.

See readme for how the game stores save game data and how it's decoded +
edited.