Small fixes and removed unused code
Fixed an issue where opening multiple files in a row would result in the
players name being appended on the end of the last file that was opened.

Changed the starting offset for the stats (the strength byte) into a
const instead of hard coded value before writing.

Removed all commented out/unused functions from an earlier version of
this program.